renaming common folder
