+++ src/tracer/src/client/events/mod.rs
@@ -2,8 +2,8 @@ mod run_details;
 use crate::cloud_providers::aws::aws_metadata::get_aws_instance_metadata;
 use crate::cloud_providers::aws::pricing::PricingSource;
 use crate::cloud_providers::aws::types::pricing::EC2FilterBuilder;
-use crate::common::debug_log::Logger;
-use crate::common::types::event::attributes::system_metrics::SystemProperties;
+use crate::process_identification::debug_log::Logger;
+use crate::process_identification::types::event::attributes::system_metrics::SystemProperties;
 use crate::extracts::metrics::system_metrics_collector::SystemMetricsCollector;
 use anyhow::Result;
 use chrono::Utc;
