+++ src/tracer/src/client/exporters/client_export_manager.rs
@@ -1,8 +1,8 @@
 use crate::client::exporters::log_writer::LogWriterEnum;
 
 use crate::client::exporters::log_writer::LogWriter;
-use crate::common::types::current_run::PipelineMetadata;
-use crate::common::types::event::Event;
+use crate::process_identification::types::current_run::PipelineMetadata;
+use crate::process_identification::types::event::Event;
 use std::sync::Arc;
 use tokio::sync::mpsc::Receiver;
 use tokio::sync::Mutex;
