+++ src/tracer/src/extracts/process/process_manager/matcher.rs
@@ -1,5 +1,5 @@
-use crate::common::target_process::target::Target;
-use crate::common::utils::log_matched_process;
+use crate::process_identification::target_process::target::Target;
+use crate::process_identification::utils::log_matched_process;
 use crate::extracts::process::types::process_state::ProcessState;
 use anyhow::Result;
 use std::collections::{HashMap, HashSet};
