+++ src/tracer/src/extracts/process/process_utils.rs
@@ -1,4 +1,4 @@
-use crate::common::types::event::attributes::process::{FullProcessProperties, ProcessProperties};
+use crate::process_identification::types::event::attributes::process::{FullProcessProperties, ProcessProperties};
 use chrono::Utc;
 use itertools::Itertools;
 use shlex;
