+++ src/tracer/src/logging.rs
@@ -1,4 +1,4 @@
-use crate::common::constants::{LOG_FILE, WORKING_DIR};
+use crate::process_identification::constants::{LOG_FILE, WORKING_DIR};
 use anyhow::{Context, Result};
 use tracing_appender::rolling::{RollingFileAppender, Rotation};
 use tracing_subscriber::{
