+++ src/tracer/src/nondaemon_commands.rs
@@ -3,7 +3,7 @@ use std::process::Command;
 #[cfg(target_os = "linux")]
 use crate::utils::system_info::get_kernel_version;
 
-use crate::common::constants::{FILE_CACHE_DIR, PID_FILE, STDERR_FILE, STDOUT_FILE};
+use crate::process_identification::constants::{FILE_CACHE_DIR, PID_FILE, STDERR_FILE, STDOUT_FILE};
 use crate::config::Config;
 use crate::daemon::client::DaemonClient;
 use crate::utils::InfoFormatter;
