+++ src/tracer/src/process_identification/debug_log/mod.rs
@@ -1,4 +1,4 @@
-use crate::common::constants::DEBUG_LOG;
+use crate::process_identification::constants::DEBUG_LOG;
 use anyhow::Context;
 use chrono::Utc;
 use serde_json::Value;
