+++ src/tracer/src/process_identification/types/event/attributes/syslog.rs
