+++ src/tracer/tests/pipeline-nfcore-rnaseq.rs
@@ -8,13 +8,13 @@ use std::sync::Arc;
 use tokio::runtime::Runtime;
 use tokio::sync::mpsc::{self, Sender};
 use tokio::sync::RwLock;
-use tracer::common::recorder::LogRecorder;
-use tracer::common::target_process::target_manager::TargetManager;
-use tracer::common::types::current_run::{PipelineMetadata, Run};
-use tracer::common::types::event::attributes::process::ProcessProperties;
-use tracer::common::types::event::attributes::EventAttributes;
-use tracer::common::types::event::{Event, ProcessStatus};
-use tracer::common::types::pipeline_tags::PipelineTags;
+use tracer::process_identification::recorder::LogRecorder;
+use tracer::process_identification::target_process::target_manager::TargetManager;
+use tracer::process_identification::types::current_run::{PipelineMetadata, Run};
+use tracer::process_identification::types::event::attributes::process::ProcessProperties;
+use tracer::process_identification::types::event::attributes::EventAttributes;
+use tracer::process_identification::types::event::{Event, ProcessStatus};
+use tracer::process_identification::types::pipeline_tags::PipelineTags;
 use tracer::extracts::ebpf_watcher::watcher::EbpfWatcher;
 use tracer_ebpf::ebpf_trigger::Trigger;
